+++ docs/conf.py
@@ -19,14 +19,18 @@ sys.path.insert(0, os.path.abspath('..'))
 
 # -- Project information -----------------------------------------------------
 
-project = 'Penman'
 copyright = '2019, Michael Wayne Goodman'
-author = 'Michael Wayne Goodman'
+
+from penman.__about__ import (
+    __version_info__ as ver,
+    __author__ as author,
+    __title__ as project,
+)
 
 # The short X.Y version
-version = 'v0.7'
+version = f'{ver[0]}.{ver[1]}'
 # The full version, including alpha/beta/rc tags
-release = 'v0.7.2'
+release = '.'.join(ver)
 
 
 # -- General configuration ---------------------------------------------------
@@ -43,6 +47,7 @@ extensions = [
     'sphinx.ext.doctest',
     'sphinx.ext.intersphinx',
     'sphinx.ext.coverage',
+    'sphinxcontrib.tikz',
     'sphinx.ext.viewcode',
     'sphinx.ext.napoleon',
 ]
@@ -176,3 +181,8 @@ intersphinx_mapping = {
 # disable type hints
 
 autodoc_typehints = 'none'
+
+
+# -- Options for SphinxContrib-Tikz ------------------------------------------
+
+tikz_proc_suite = 'GhostScript'