+++ CHANGELOG.md
@@ -2,9 +2,25 @@
 
 ## [Unreleased][unreleased]
 
-(No unreleased changes)
+(no unreleased changes yet)
 
-## [v0.7.2][v0.7.2]
+## [v0.8.0][]
+
+This release adds a handful of minor functions and expands the
+documentation.
+
+### Added
+
+* `penman.layout.appears_inverted()` ([#37][])
+* Tree branch sorting functions ([#31][])
+  - `penman.model.Model.original_order()`
+  - `penman.model.Model.canonical_order()`
+  - `penman.model.Model.random_order()`
+* `penman.layout.rearrange()` ([#31][])
+* `--rearrange=(canonical|random)` command-line option
+
+
+## [v0.7.2][]
 
 ### Added
 
@@ -25,7 +41,7 @@
 * `Model.reify()` no longer inverts the incoming triple
 
 
-## [v0.7.1][v0.7.1]
+## [v0.7.1][]
 
 The [v0.7.0](#v070) release was missing a declaration for the new
 `penman.models` sub-package, so it was not available for new
@@ -39,7 +55,7 @@ installs. This release fixes that.
   culprit for faulty documentation builds.)
 
 
-## [v0.7.0][v0.7.0]
+## [v0.7.0][]
 
 This release comprises a major restructuring from previous
 versions. No longer is there a single `penman.py` module, but the
@@ -198,7 +214,7 @@ modules.
   reentrancy issues ([#29][])
 
 
-## [v0.6.2][v0.6.2]
+## [v0.6.2][]
 
 ### Fixed
 
@@ -209,7 +225,7 @@ modules.
   default parser (the numeric values are still parses as atoms) ([#17][])
 
 
-## [v0.6.1][v0.6.1]
+## [v0.6.1][]
 
 ### Added
 
@@ -411,6 +427,7 @@ First release with very basic functionality.
 [v0.7.0]: ../../releases/tag/v0.7.0
 [v0.7.1]: ../../releases/tag/v0.7.1
 [v0.7.2]: ../../releases/tag/v0.7.2
+[v0.8.0]: ../../releases/tag/v0.8.0
 [README]: README.md
 
 [#4]: https://github.com/goodmami/penman/issues/4
@@ -430,6 +447,8 @@ First release with very basic functionality.
 [#26]: https://github.com/goodmami/penman/issues/26
 [#27]: https://github.com/goodmami/penman/issues/27
 [#29]: https://github.com/goodmami/penman/issues/29
+[#31]: https://github.com/goodmami/penman/issues/31
 [#32]: https://github.com/goodmami/penman/issues/32
 [#34]: https://github.com/goodmami/penman/issues/34
 [#35]: https://github.com/goodmami/penman/issues/35
+[#37]: https://github.com/goodmami/penman/issues/37